Fix reporting of bogus WiFi config.
Looks like the {station,softap}_config structs changed at some point and
we didn't pick up on it. Now paying attention to the return value of the
SDK functions and only reporting config if we actually got some config
back.

Also, zero-initialise the config structs before use to be on the safe side.